+++ example/lib/custom_animation.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
+
+class CustomAnimation extends EasyLoadingAnimation {
+  CustomAnimation();
+
+  @override
+  Widget buildWidget(
+    Widget child,
+    AnimationController controller,
+    AlignmentGeometry alignment,
+  ) {
+    double opacity = controller?.value ?? 0;
+    return Opacity(
+      opacity: opacity,
+      child: RotationTransition(
+        turns: controller,
+        child: child,
+      ),
+    );
+  }
+}