+++ lib/src/animations/animation.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/widgets.dart';
+
+abstract class EasyLoadingAnimation {
+  EasyLoadingAnimation();
+
+  Widget call(
+    Widget child,
+    AnimationController controller,
+    AlignmentGeometry alignment,
+  ) {
+    return buildWidget(
+      child,
+      controller,
+      alignment,
+    );
+  }
+
+  Widget buildWidget(
+    Widget child,
+    AnimationController controller,
+    AlignmentGeometry alignment,
+  );
+}