+++ lib/src/easy_loading.dart
@@ -1,16 +1,44 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 
 import './widgets/container.dart';
 import './widgets/progress.dart';
 import './widgets/indicator.dart';
+import './widgets/overlay_entry.dart';
+import './animations/animation.dart';
 import './theme.dart';
 
 /// loading style
 enum EasyLoadingStyle {
   light,
   dark,
+  custom,
+}
+
+/// toast position
+enum EasyLoadingToastPosition {
+  top,
+  center,
+  bottom,
+}
+
+/// loading animation
+enum EasyLoadingAnimationStyle {
+  opacity,
+  offset,
+  scale,
+  custom,
+}
+
+/// loading mask type
+/// [none] default mask type, allow user interactions while loading is displayed
+/// [clear] don't allow user interactions while loading is displayed
+/// [black] don't allow user interactions while loading is displayed
+/// [custom] while mask type is custom, maskColor should not be null
+enum EasyLoadingMaskType {
+  none,
+  clear,
+  black,
   custom,
 }
 
@@ -41,18 +69,6 @@ enum EasyLoadingIndicatorType {
   squareCircle,
 }
 
-/// loading mask type
-/// [none] default mask type, allow user interactions while loading is displayed
-/// [clear] don't allow user interactions while loading is displayed
-/// [black] don't allow user interactions while loading is displayed
-/// [custom] while mask type is custom, maskColor should not be null
-enum EasyLoadingMaskType {
-  none,
-  clear,
-  black,
-  custom,
-}
-
 class EasyLoading {
   /// loading style, default [EasyLoadingStyle.dark].
   EasyLoadingStyle loadingStyle;
@@ -62,6 +78,15 @@ class EasyLoading {
 
   /// loading mask type, default [EasyLoadingMaskType.none].
   EasyLoadingMaskType maskType;
+
+  /// toast position, default [EasyLoadingToastPosition.center].
+  EasyLoadingToastPosition toastPosition;
+
+  /// loading animationStyle, default [EasyLoadingAnimationStyle.opacity].
+  EasyLoadingAnimationStyle animationStyle;
+
+  /// loading custom animation, default null.
+  EasyLoadingAnimation customAnimation;
 
   /// textAlign of status, default [TextAlign.center].
   TextAlign textAlign;
@@ -93,6 +118,9 @@ class EasyLoading {
   /// display duration of [showSuccess] [showError] [showInfo] [showToast], default 2000ms.
   Duration displayDuration;
 
+  /// animation duration of indicator, default 200ms.
+  Duration animationDuration;
+
   /// color of loading status, only used for [EasyLoadingStyle.custom].
   Color textColor;
 
@@ -123,16 +151,16 @@ class EasyLoading {
   /// info widget of loading
   Widget infoWidget;
 
-  OverlayEntry overlayEntry;
+  EasyLoadingOverlayEntry overlayEntry;
   Widget _w;
 
-  GlobalKey<LoadingContainerState> _key;
-  GlobalKey<ProgressState> _progressKey;
+  GlobalKey<EasyLoadingContainerState> _key;
+  GlobalKey<EasyLoadingProgressState> _progressKey;
   Timer _timer;
 
   Widget get w => _w;
-  GlobalKey<LoadingContainerState> get key => _key;
-  GlobalKey<ProgressState> get progressKey => _progressKey;
+  GlobalKey<EasyLoadingContainerState> get key => _key;
+  GlobalKey<EasyLoadingProgressState> get progressKey => _progressKey;
   Timer get timer => _timer;
 
   factory EasyLoading() => _getInstance();
@@ -144,6 +172,8 @@ class EasyLoading {
     loadingStyle = EasyLoadingStyle.dark;
     indicatorType = EasyLoadingIndicatorType.fadingCircle;
     maskType = EasyLoadingMaskType.none;
+    toastPosition = EasyLoadingToastPosition.center;
+    animationStyle = EasyLoadingAnimationStyle.opacity;
     textAlign = TextAlign.center;
     indicatorSize = 40.0;
     radius = 5.0;
@@ -151,6 +181,7 @@ class EasyLoading {
     progressWidth = 2.0;
     lineWidth = 4.0;
     displayDuration = const Duration(milliseconds: 2000);
+    animationDuration = const Duration(milliseconds: 200);
     textPadding = const EdgeInsets.only(bottom: 10.0);
     contentPadding = const EdgeInsets.symmetric(
       vertical: 15.0,
@@ -189,8 +220,9 @@ class EasyLoading {
       'progress value should be 0.0 ~ 1.0',
     );
     if (_getInstance()._progressKey == null || _getInstance().w == null) {
-      GlobalKey<ProgressState> _progressKey = GlobalKey<ProgressState>();
-      Widget w = Progress(
+      GlobalKey<EasyLoadingProgressState> _progressKey =
+          GlobalKey<EasyLoadingProgressState>();
+      Widget w = EasyLoadingProgress(
         key: _progressKey,
         value: value,
       );
@@ -265,10 +297,12 @@ class EasyLoading {
   static void showToast(
     String status, {
     Duration duration,
+    EasyLoadingToastPosition toastPosition,
   }) {
     _getInstance()._show(
       status: status,
       duration: duration ?? EasyLoadingTheme.displayDuration,
+      toastPosition: toastPosition ?? EasyLoadingTheme.toastPosition,
     );
   }
 
@@ -278,14 +312,7 @@ class EasyLoading {
   }) async {
     // cancel timer
     _getInstance()._cancelTimer();
-    if (SchedulerBinding.instance.schedulerPhase ==
-        SchedulerPhase.persistentCallbacks) {
-      SchedulerBinding.instance.addPostFrameCallback((_) {
-        _getInstance()._dismiss(animation);
-      });
-    } else {
-      _getInstance()._dismiss(animation);
-    }
+    _getInstance()._dismiss(animation);
   }
 
   /// show loading
@@ -293,10 +320,16 @@ class EasyLoading {
     Widget w,
     String status,
     Duration duration,
+    EasyLoadingToastPosition toastPosition = EasyLoadingToastPosition.center,
   }) {
     assert(
       _getInstance().overlayEntry != null,
       'overlayEntry should not be null',
+    );
+
+    assert(
+      toastPosition != null,
+      'toastPosition should not be null',
     );
 
     if (_getInstance().loadingStyle == EasyLoadingStyle.custom) {
@@ -325,15 +358,24 @@ class EasyLoading {
       );
     }
 
+    if (_getInstance().animationStyle == EasyLoadingAnimationStyle.custom) {
+      assert(
+        _getInstance().customAnimation != null,
+        'while animationStyle is custom, customAnimation should not be null',
+      );
+    }
+
     _cancelTimer();
     _getInstance()._progressKey = null;
 
-    GlobalKey<LoadingContainerState> _key = GlobalKey<LoadingContainerState>();
-    _getInstance()._w = LoadingContainer(
+    GlobalKey<EasyLoadingContainerState> _key =
+        GlobalKey<EasyLoadingContainerState>();
+    _getInstance()._w = EasyLoadingContainer(
       key: _key,
       status: status,
       indicator: w,
       animation: _getInstance()._w == null,
+      toastPosition: toastPosition,
     );
     _markNeedsBuild();
     _getInstance()._key = _key;
@@ -346,11 +388,11 @@ class EasyLoading {
 
   void _dismiss(bool animation) {
     if (animation) {
-      LoadingContainerState loadingContainerState =
+      EasyLoadingContainerState easyLoadingContainerState =
           _getInstance().key?.currentState;
-      if (loadingContainerState != null) {
+      if (easyLoadingContainerState != null) {
         final Completer<void> completer = Completer<void>();
-        loadingContainerState.dismiss(completer);
+        easyLoadingContainerState.dismiss(completer);
         completer.future.then((value) {
           _reset();
         });
@@ -368,14 +410,7 @@ class EasyLoading {
   }
 
   void _markNeedsBuild() {
-    if (SchedulerBinding.instance.schedulerPhase ==
-        SchedulerPhase.persistentCallbacks) {
-      SchedulerBinding.instance.addPostFrameCallback((_) {
-        _getInstance().overlayEntry?.markNeedsBuild();
-      });
-    } else {
-      _getInstance().overlayEntry?.markNeedsBuild();
-    }
+    _getInstance().overlayEntry?.markNeedsBuild();
   }
 
   void _cancelTimer() {