+++ lib/src/widgets/container.dart
@@ -2,60 +2,56 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import '../theme.dart';
+import '../easy_loading.dart';
 
-class LoadingContainer extends StatefulWidget {
+class EasyLoadingContainer extends StatefulWidget {
   final Widget indicator;
   final String status;
   final bool animation;
+  final EasyLoadingToastPosition toastPosition;
 
-  LoadingContainer({
+  const EasyLoadingContainer({
     Key key,
     this.indicator,
     this.status,
     this.animation = true,
+    this.toastPosition,
   }) : super(key: key);
 
   @override
-  LoadingContainerState createState() => LoadingContainerState();
+  EasyLoadingContainerState createState() => EasyLoadingContainerState();
 }
 
-class LoadingContainerState extends State<LoadingContainer> {
-  double _opacity = 0.0;
-  Duration _animationDuration;
+class EasyLoadingContainerState extends State<EasyLoadingContainer>
+    with SingleTickerProviderStateMixin {
   String _status;
+  AnimationController _animationController;
 
   @override
   void initState() {
     super.initState();
     if (!mounted) return;
     _status = widget.status;
-    _animationDuration = widget.animation
-        ? const Duration(milliseconds: 300)
-        : const Duration(milliseconds: 0);
-    if (widget.animation) {
-      Future.delayed(const Duration(milliseconds: 30), () {
-        setState(() {
-          _opacity = 1.0;
-        });
-      });
-    } else {
-      setState(() {
-        _opacity = 1.0;
-      });
-    }
+
+    _animationController = AnimationController(
+      vsync: this,
+      duration: EasyLoadingTheme.animationDuration,
+      value: widget.animation ? 0 : 1,
+    );
+    _animationController.forward();
   }
 
   @override
   void dispose() {
+    _animationController?.dispose();
+    _animationController = null;
     super.dispose();
   }
 
   void dismiss(Completer completer) {
-    _animationDuration = const Duration(milliseconds: 300);
-    setState(() {
-      _opacity = 0.0;
-    });
-    Future.delayed(_animationDuration, () {
+    _animationController?.reverse();
+    Timer.periodic(EasyLoadingTheme.animationDuration, (Timer timer) {
+      timer?.cancel();
       completer.complete();
     });
   }
@@ -69,61 +65,92 @@ class LoadingContainerState extends State<LoadingContainer> {
 
   @override
   Widget build(BuildContext context) {
-    Widget loading = Align(
-      alignment: Alignment.center,
-      child: Container(
-        margin: const EdgeInsets.all(50.0),
-        decoration: BoxDecoration(
-          color: EasyLoadingTheme.backgroundColor,
-          borderRadius: BorderRadius.circular(
-            EasyLoadingTheme.radius,
-          ),
+    AlignmentGeometry _alignment =
+        (widget.indicator == null && widget.status?.isNotEmpty == true)
+            ? EasyLoadingTheme.alignment(widget.toastPosition)
+            : AlignmentDirectional.center;
+    return Stack(
+      alignment: _alignment,
+      children: <Widget>[
+        AnimatedBuilder(
+          animation: _animationController,
+          builder: (BuildContext context, Widget child) {
+            return Opacity(
+              opacity: _animationController?.value ?? 0,
+              child: IgnorePointer(
+                ignoring: EasyLoadingTheme.ignoring,
+                child: Container(
+                  width: double.infinity,
+                  height: double.infinity,
+                  color: EasyLoadingTheme.maskColor,
+                ),
+              ),
+            );
+          },
         ),
-        padding: EasyLoadingTheme.contentPadding,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            widget.indicator != null
-                ? Container(
-                    margin: _status?.isNotEmpty == true
-                        ? EasyLoadingTheme.textPadding
-                        : EdgeInsets.zero,
-                    child: widget.indicator,
-                  )
-                : null,
-            _status?.isNotEmpty == true
-                ? Text(
-                    _status,
-                    style: EasyLoadingTheme.textStyle ??
-                        TextStyle(
-                          color: EasyLoadingTheme.textColor,
-                          fontSize: EasyLoadingTheme.fontSize,
-                        ),
-                    textAlign: EasyLoadingTheme.textAlign,
-                  )
-                : null,
-          ].where((w) => w != null).toList(),
+        AnimatedBuilder(
+          animation: _animationController,
+          builder: (BuildContext context, Widget child) {
+            return EasyLoadingTheme.loadingAnimation?.buildWidget(
+              _Indicator(
+                status: _status,
+                indicator: widget.indicator,
+              ),
+              _animationController,
+              _alignment,
+            );
+          },
+        ),
+      ],
+    );
+  }
+}
+
+class _Indicator extends StatelessWidget {
+  final Widget indicator;
+  final String status;
+
+  const _Indicator({
+    @required this.indicator,
+    @required this.status,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.all(50.0),
+      decoration: BoxDecoration(
+        color: EasyLoadingTheme.backgroundColor,
+        borderRadius: BorderRadius.circular(
+          EasyLoadingTheme.radius,
         ),
       ),
-    );
-
-    return AnimatedOpacity(
-      opacity: _opacity,
-      duration: _animationDuration,
-      child: Stack(
+      padding: EasyLoadingTheme.contentPadding,
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          IgnorePointer(
-            ignoring: EasyLoadingTheme.ignoring,
-            child: Container(
-              width: double.infinity,
-              height: double.infinity,
-              color: EasyLoadingTheme.maskColor,
-            ),
-          ),
-          loading,
-        ],
+          indicator != null
+              ? Container(
+                  margin: status?.isNotEmpty == true
+                      ? EasyLoadingTheme.textPadding
+                      : EdgeInsets.zero,
+                  child: indicator,
+                )
+              : null,
+          status?.isNotEmpty == true
+              ? Text(
+                  status,
+                  style: EasyLoadingTheme.textStyle ??
+                      TextStyle(
+                        color: EasyLoadingTheme.textColor,
+                        fontSize: EasyLoadingTheme.fontSize,
+                      ),
+                  textAlign: EasyLoadingTheme.textAlign,
+                )
+              : null,
+        ].where((w) => w != null).toList(),
       ),
     );
   }