+++ lib/src/widgets/loading.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../easy_loading.dart';
+import './overlay_entry.dart';
 
 class FlutterEasyLoading extends StatefulWidget {
   final Widget child;
@@ -8,19 +9,20 @@ class FlutterEasyLoading extends StatefulWidget {
   const FlutterEasyLoading({
     Key key,
     @required this.child,
-  }) : super(key: key);
+  })  : assert(child != null),
+        super(key: key);
 
   @override
   _FlutterEasyLoadingState createState() => _FlutterEasyLoadingState();
 }
 
 class _FlutterEasyLoadingState extends State<FlutterEasyLoading> {
-  OverlayEntry _overlayEntry;
+  EasyLoadingOverlayEntry _overlayEntry;
 
   @override
   void initState() {
     super.initState();
-    _overlayEntry = OverlayEntry(
+    _overlayEntry = EasyLoadingOverlayEntry(
       builder: (BuildContext context) => EasyLoading.instance.w ?? Container(),
     );
     EasyLoading.instance.overlayEntry = _overlayEntry;
@@ -37,7 +39,9 @@ class _FlutterEasyLoadingState extends State<FlutterEasyLoading> {
     return Material(
       child: Overlay(
         initialEntries: [
-          OverlayEntry(builder: (BuildContext context) => widget.child),
+          EasyLoadingOverlayEntry(
+            builder: (BuildContext context) => widget.child,
+          ),
           _overlayEntry,
         ],
       ),